requirements.txt
torch>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl
tqdm>=4.65.0
Pillow>=9.0.0